Cleanup of NSIS files.
Results of the compilation process should not be part of the repository.
Thus BUSMASTER_Cleanup_Registry.exe was removed. It can be generated
by compiling CleanRegistry.msi. This is done now by the following...

The build.bat has been extended to cope with 64-bit Windows versions.
The compilation of CleanRegistry.nsi was build into the script to
generate BUSMASTER_Cleanup_Registry.exe automatically.

Signed-off-by: Tobias Lorenz <tobias.lorenz@etas.com>