Fix transmission of signals by Test Automation Executor
The order of signals obtained by pMsg->GetSignalList( signalList ); is
determined by keys of the map. In this case it is ISignal* and
having items sorted on the address of a pointer makes no sense.
Moreover on every invocation of GetSignalList() we get another map of
pointers so perhaps there is a memory leak as well.

The order of signals in ouSendData.m_odSignalDataList is determined by the signals
present in the testsetup xml file and we cannot make any assumptions on this.

So the signals in the signalList (containing startaddresses, endianess etc)
must be matched to the ouSendData.m_odSignalDataList on signalName.

The GetProperties (DBManager.dll ?) function previously used in un64GetBitMask
did not initialize m_ouEndianess and this results in unpredictable results.
For CAN this was the last place in BUSMASTER m_ouEndianess was used.